+++ src/errors/AbortedException.php
@@ -1,0 +1,13 @@
+<?php
+namespace VovanVE\parser\errors;
+
+use VovanVE\parser\SyntaxException;
+
+/**
+ * Developer-defined error from aborted action
+ * @package VovanVE\parser
+ * @since 1.6.0
+ */
+class AbortedException extends SyntaxException
+{
+}